+++ packages/form-render/src/form-items/render/index.tsx
@@ -1,11 +1,11 @@
 import type { DataRecord } from '../../interfaces'
 
-export function renderRenderItem(renderer: RenderInputItemOptions) {
-  return (data: DataRecord) => {
+export function renderRenderItem<T=DataRecord>(renderer: RenderInputItemOptions<T>) {
+  return (data: T) => {
     return renderer(data)
   }
 }
 
-export interface RenderInputItemOptions {
-  (record: any): JSX.Element
+export interface RenderInputItemOptions<T> {
+  (record: T): JSX.Element
 }