+++ packages/form-render/src/form-items/textarea/index.tsx
@@ -1,11 +1,11 @@
 import { Textarea } from '@arco-design/web-vue'
 import type { DataRecord, FormItemOptions } from '../../interfaces'
 
-export function renderTextareaItem(options?: RenderTextareaItemOptions) {
-  return (data: DataRecord, form: FormItemOptions) => {
+export function renderTextareaItem<T=DataRecord>(options?: RenderTextareaItemOptions) {
+  return (data: T, form: FormItemOptions<T>) => {
     return (
       <Textarea
-        v-model={data[form.key]}
+        v-model={data[form.key as keyof T]}
         placeholder={options?.placeholder}
         autoSize={options?.autosize}
         maxLength={options?.maxLength}></Textarea>