+++ packages/form-render/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gopowerteam/form-render
 
+## 0.0.6
+
+### Patch Changes
+
+- update d.ts
+
+## 0.0.5
+
+### Patch Changes
+
+- update d.ts
+
 ## 0.0.4
 
 ### Patch Changes